fix(same-props): fix bug where sameProps depended on chronological ordering of keys
